Move implicit conversions to the syntax object
